Style : " GameObjectInfo 수정 "
GameServer
 - GameObjectInfo
  - st_StatInfo 수정
   - Level ( 자료형 int32 -> int8 ) 수정

+++ GameServer/GameObjectInfo.h
@@ -1045,7 +1045,7 @@ struct st_PositionInfo
 
 struct st_StatInfo
 {
-	int32 Level;
+	int8 Level;
 	int32 Str;
 	int32 Dex;
 	int32 Int;
